requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
akshare>=1.14.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
requests>=2.31.0
aiohttp>=3.9.1
jinja2>=3.1.3
apscheduler>=3.10.4
psycopg2-binary>=2.9.9
asyncpg>=0.29.0